+++ CHANGES.md
@@ -3,6 +3,13 @@ Pykit Changelog
 
 Here you can see the full list of changes between each Pykit release.
 
+Version 0.3.10
+-----------
+- changed: config `scroll` option `x` changed to `xy`
+- changed: config `valign` option `top` changed to `parent`
+- added: help text, and cleaner form control display
+- upgrade uikit version to 2.26.x
+
 Version 0.3.9
 -----------
 - added: generic positioning methods to `dropdown`